refactor: consolidate KOTH into features/king-of-the-hill module
