+++ src/features/king-of-the-hill/components/GradientBorderContent.tsx
