+++ src/features/king-of-the-hill/components/HeaderButton.tsx
@@ -2,10 +2,11 @@ import React from 'react';
 import { StyleSheet, View } from 'react-native';
 
 import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
-import { CHEVRON_RIGHT_SYMBOL } from '@/components/king-of-the-hill/constants';
-import { GradientBorderContent } from '@/components/king-of-the-hill/GradientBorderContent';
 import { RainbowImage } from '@/components/RainbowImage';
 import { Text } from '@/design-system';
+
+import { CHEVRON_RIGHT_SYMBOL } from '../constants';
+import { GradientBorderContent } from './GradientBorderContent';
 
 type HeaderButtonProps = {
   onPress: () => void;
