+++ src/features/king-of-the-hill/components/LaunchButton.tsx
