+++ src/features/king-of-the-hill/components/RainbowGlow.tsx
