+++ src/features/king-of-the-hill/constants.ts
