+++ src/features/king-of-the-hill/hooks/useShowKingOfTheHill.ts
