+++ src/features/king-of-the-hill/stores/kingOfTheHillStore.ts
@@ -2,11 +2,10 @@ import { metadataClient } from '@/graphql';
 import { type KingOfTheHill, type KingOfTheHillRankings } from '@/graphql/__generated__/metadata';
 import { logger, RainbowError } from '@/logger';
 import Routes from '@/navigation/routesNames';
+import { userAssetsStoreManager } from '@/state/assets/userAssetsStoreManager';
 import { createQueryStore } from '@/state/internal/createQueryStore';
+import { useNavigationStore } from '@/state/navigation/navigationStore';
 import { time } from '@/utils/time';
-
-import { userAssetsStoreManager } from '../assets/userAssetsStoreManager';
-import { useNavigationStore } from '../navigation/navigationStore';
 
 type State = {
   kingOfTheHill?: KingOfTheHill;
